# openapi.yaml
openapi: 3.1.0
info:
  title: ClickUp Bridge
  version: "1.1.0"
servers:
  - url: https://clickup.zynodo.com
paths:
  /health:
    get:
      operationId: health
      responses:
        "200": { description: OK }
  /api/team:
    get:
      operationId: listTeams
      responses:
        "200": { description: Teams JSON }
  /api/team/{teamId}/space:
    get:
      operationId: listSpaces
      parameters:
        - { name: teamId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Spaces JSON }
  /api/space/{spaceId}/list:
    get:
      operationId: listLists
      parameters:
        - { name: spaceId, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: Lists JSON }
  /api/list/{listId}/task:
    post:
      operationId: createTask
      parameters:
        - { name: listId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                assignees: { type: array, items: { type: integer } }
              required: [name]
      responses:
        "200": { description: Task JSON }
  /commands/find_space:
    get:
      operationId: findSpace
      parameters:
        - { name: teamId, in: query, required: true, schema: { type: string } }
        - { name: name, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: Matched spaces }
  /commands/find_folder:
    get:
      operationId: findFolder
      parameters:
        - { name: spaceId, in: query, required: true, schema: { type: string } }
        - { name: name, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: Matched folders }
  /commands/find_list:
    get:
      operationId: findList
      parameters:
        - { name: spaceId, in: query, required: false, schema: { type: string } }
        - { name: folderId, in: query, required: false, schema: { type: string } }
        - { name: name, in: query, required: true, schema: { type: string } }
      responses:
        "200": { description: Matched lists }
  /commands/search_tasks:
    get:
      operationId: searchTasks
      parameters:
        - { name: listId, in: query, required: true, schema: { type: string } }
        - { name: nameContains, in: query, required: false, schema: { type: string } }
        - { name: updatedFrom, in: query, required: false, schema: { type: string, description: "ISO or epoch ms" } }
        - { name: updatedTo, in: query, required: false, schema: { type: string, description: "ISO or epoch ms" } }
        - { name: assignee, in: query, required: false, schema: { type: string } }
        - { name: status, in: query, required: false, schema: { type: string } }
        - { name: page, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer, default: 100 } }
      responses:
        "200": { description: Tasks list }
  /commands/task_comments:
    get:
      operationId: taskComments
      parameters:
        - { name: taskId, in: query, required: true, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer } }
      responses:
        "200": { description: Comments list }
  /commands/create_task:
    post:
      operationId: cmdCreateTask
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                listId: { type: string }
                name: { type: string }
                description: { type: string }
                assignees: { type: array, items: { type: integer } }
              required: [listId, name]
      responses:
        "200": { description: Created task }
components:
  schemas: {}
security:
  - {}